+++ www/bluetoothle.js
@@ -1,7 +1,7 @@
 var bluetoothleName = "BluetoothLePlugin";
 var bluetoothle = {
-  initialize: function(successCallback, errorCallback, params) {
-    cordova.exec(successCallback, errorCallback, bluetoothleName, "initialize", [params]);
+  initialize: function(successCallback, params) {
+    cordova.exec(successCallback, successCallback, bluetoothleName, "initialize", [params]);
   },
   enable: function(successCallback, errorCallback) {
     cordova.exec(successCallback, errorCallback, bluetoothleName, "enable", []);
@@ -84,11 +84,11 @@ var bluetoothle = {
   requestConnectionPriority: function(successCallback, errorCallback, params) {
     cordova.exec(successCallback, errorCallback, bluetoothleName, "requestConnectionPriority", [params]);
   },
-  hasPermission: function(successCallback) {
-    cordova.exec(successCallback, successCallback, bluetoothleName, "hasPermission", []);
+  hasPermission: function(successCallback, errorCallback) {
+    cordova.exec(successCallback, errorCallback, bluetoothleName, "hasPermission", []);
   },
-  requestPermission: function(successCallback) {
-    cordova.exec(successCallback, successCallback, bluetoothleName, "requestPermission", []);
+  requestPermission: function(successCallback, errorCallback) {
+    cordova.exec(successCallback, errorCallback, bluetoothleName, "requestPermission", []);
   },
   encodedStringToBytes: function(string) {
     var data = atob(string);
@@ -112,6 +112,18 @@ var bluetoothle = {
   },
   bytesToString: function(bytes) {
     return String.fromCharCode.apply(null, new Uint16Array(bytes));
-  }
+  },
+  SCAN_MODE_OPPORTUNISTIC: -1,
+  SCAN_MODE_LOW_POWER: 0,
+  SCAN_MODE_BALANCED: 1,
+  SCAN_MODE_LOW_LATENCY: 2,
+  MATCH_NUM_ONE_ADVERTISEMENT: 1,
+  MATCH_NUM_FEW_ADVERTISEMENT: 2,
+  MATCH_NUM_MAX_ADVERTISEMENT: 3,
+  MATCH_MODE_AGGRESSIVE: 1,
+  MATCH_MODE_STICKY: 2,
+  CALLBACK_TYPE_ALL_MATCHES: 1,
+  CALLBACK_TYPE_FIRST_MATCH: 2,
+  CALLBACK_TYPE_MATCH_LOST: 4,
 }
 module.exports = bluetoothle;